safari-600.7-branch/LayoutTests/platform/mac/accessibility/input-title.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<title>Input Title</title>
</head>
<body>

<input id="green" name="color" title="green" type="radio">
<input id="red" name="color" title="red" type="radio">
<label for="red"></label>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that the input title is exposed correctly.");
    
    if (window.accessibilityController) {
        var green = accessibilityController.accessibleElementById("green");
        shouldBe("green.description", "'AXDescription: green'");
        
        var red = accessibilityController.accessibleElementById("red");
        shouldBe("red.description", "'AXDescription: red'");
    }
</script>

<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
